+++ src/components/utils/helpApi.js
@@ -1,8 +1,22 @@
 import { api, getHeader, getToken, getHeaderForFormData } from "./AppFunction";
 
+const isHelpCategoryValid = ["USER","BUYER","PRODUCT","ORDER","INVENTORY","SHIFT","GENERAL",
+    "SYSTEM_SETTING","PROCUREMENT","GOODS","RAW_MATERIAL","INVOICE","PAYMENT","ROLE","SALES","VENDOR","SHELF",
+    "CONFIRMATION_DOCUMENT","ITEM_SALES","SALES_ORDER","STORAGE","SUPPLY","TOKEN","SUPPLY_ITEM","INVENTORY_ITEMS","SHIFT_REPORT","BARCODE","VEHICLE",
+    "DRIVER","ROUTE","SHIPMENT","LOGISTICS_PROVIDER","STOCK_TRANSFER","STOCK_TRANSFER_ITEM","TRACKING_INFO","TRANSPORT_ORDER","OUTBOUND_DELIVERY",
+    "INBOUND_DELIVERY","DELIVERY_ITEM","COMPANY_EMAIL","DASHBOARD","OPTION","EMAI_SETTING"
+]; 
 
 export async function createHelp(title,content,category,isVisible){
     try{
+        if(
+            !title || typeof title !== "string" || title.trim() === ""||
+            !content || typeof content !== "string" || content.trim() === ""||
+            !isHelpCategoryValid.includes(category.toUpperCase()) ||
+            typeof isVisible !== "boolean"
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {title,content,category: (category || "").toUpperCase(),isVisible};
         const response = await api.post(`${import.meta.env.VITE_API_BASE_URL}/help/create/`,requestBody,{
             headers:getHeader()
@@ -21,6 +35,15 @@ export async function createHelp(title,content,category,isVisible){
 
 export async function updateHelp(id,title,content,category,isVisible){
     try{
+        if(
+            !id ||
+            !title || typeof title !== "string" || title.trim() === ""||
+            !content || typeof content !== "string" || content.trim() === ""||
+            !isHelpCategoryValid.includes(category.toUpperCase()) ||
+            typeof isVisible !== "boolean"
+        ){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
         const requestBody = {title,content,category: (category || "").toUpperCase(),isVisible};
         const response = await api.put(`${import.meta.env.VITE_API_BASE_URL}/help/update/${id}`,requestBody,{
             headers:getHeader()
@@ -39,6 +62,9 @@ export async function updateHelp(id,title,content,category,isVisible){
 
 export async function deleteHelp(id){
     try{
+        if(!id){
+            throw new Error("Dati ID za help nije pronadjen");
+        }
         const response = await api.delete(`${import.meta.env.VITE_API_BASE_URL}/help/delete/${id}`,{
             headers:getHeader()
         });
@@ -51,6 +77,9 @@ export async function deleteHelp(id){
 
 export async function getById(id){
     try{
+        if(!id){
+            throw new Error("Dati ID za help nije pronadjen");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/help/get-one/${id}`,{
             headers:getHeader()
         });
@@ -87,6 +116,9 @@ export async function getVisible(){
 
 export async function getByCategory(category){
     try{
+        if(!isHelpCategoryValid.includes(category.toUpperCase())){
+            throw new Error("Nepostojeca kategorija");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/help/category/${category}`,{
             headers:getHeader()
         });
